f/src/components/controllers/PlayBar: add selection drag and coloring
If a playbar bound is right clicked, the opposite bound will change
position as well.

Additionally, add coloring to reflect the areas of the playbar that
are off limits. The user will still be able to position the playpos
out of bounds, and this background coloring will need to be reworked
once a reliable playpos interval handler is developed.