f/src/componets/player/player: add position control and setters/getters
The tracks position can now be tracked fairly accurately, but there are
still improvements to be made. Player should be passed a position
callback to be called at regularly scheduled intervals, and also
irregularly when the player stops and starts and loops and changes
direction etc.

Additionally, convert the enabler and changer functions to setters and
getters as they are more readable.