f/src: grow TrackManipulator and MashupVisualizer
TrackManipulator will store multiple plugins that cannot all fit on the
screen at once. It should take up the remaining space in the column.

MashupVisualizer is the most intriguing part of the screen and should
also take up as much space as possible.

Also prevent users from accessing the context menu when right clicking
in the Track columns, as these have specific right click behavior and
it is annoying when a context menu accidentally pops up.