f/src/components/controllers/PlayBar: do not edit boundaries on reverse
Changing boundary color and tooltip seemed like a cool idea at first,
but it makes things more confusing. Too much changes on the screen at
once, especially when the playbar is active.

Reverse means "play from end", not "play from start which used to be the
end but is now start."

I do think the playbar should still anchor to the end when reversing
though, it makes it easier to see how much of the track is left and
doesn't fill up the entire playbar when it is stopped.